Enlightens BasicBlockModel and SimpleBlockcModel to ParallelInstructi… (#3)
* Enlightens BasicBlockModel and SimpleBlockcModel to ParallelInstructionLanguageHelper

Previously BasicBlockModel and SimpleBlockModel might terminate a block
in the middle of an instruction group

Now BasicBlockModel and SimpleBlockModel respect instruction group
boundaries

Adds getFlowType to ParallelInstructionLanguageHelper and implements it
in HexagonParallelInstructionLanguageHelper

Reverts formatting on ParallelInstructionLanguageHelper for a cleaner
diff off of master

Adds tests, HexagonPacketTestCodeBlock

As of this commit, the function graph also respects instruction group
boundaries, so there are no longer awkward flows in between packets. For
example:

    {  SL2_jumpr31
       SA1_seti  R0 0x2 }

Previously appeared as

    SL2_jumpr31 -> SA1_seti R0 0x2

In the function graph view. These are now displayed as one basic block.

* Adds Parallel fields by default for Function Graph

These fields are cached in the ~/.ghidra/.ghidra_10.2_DEV/ folder so you
will need to delete that directory or manually enable those fields to
take effect

* Updates README